// src/models/PokemonObject.tsx
export interface PokemonObject {
    abilities: AbilityState[];
    base_experience: number;
    cries: Cries;
    forms: Ability[];
    game_indices: Gameindex[];
    height: number;
    held_items: any[];
    id: number;
    is_default: boolean;
    location_area_encounters: string;
    moves: Move[];
    name: string;
    order: number;
    past_abilities: any[];
    past_types: any[];
    species: Ability;
    sprites: Sprites;
    stats: Stat[];
    types: Type[];
    weight: number;
  }
  interface Type {
    slot: number;
    type: Ability;
  }
  interface Stat {
    base_stat: number;
    effort: number;
    stat: Ability;
  }
  interface Sprites {
    back_default: string;
    back_female?: any;
    back_shiny: string;
    back_shiny_female?: any;
    front_default: string; // this is Image
    front_female?: any;
    front_shiny: string;
    front_shiny_female?: any;
    other: Other;
    versions: Versions;
  }
  interface Versions {
    'generation-i': Generationi;
    'generation-ii': Generationii;
    'generation-iii': Generationiii;
    'generation-iv': Generationiv;
    'generation-v': Generationv;
    'generation-vi': Generationvi;
    'generation-vii': Generationvii;
    'generation-viii': Generationviii;
  }
// ...
  interface Other {
    dream_world: Dreamworld;
    home: Home;
    'official-artwork': Officialartwork;
    showdown: Showdown;
  }
// ...
  interface Move {
    move: Ability;
    version_group_details: Versiongroupdetail[];
  }
  interface Versiongroupdetail {
    level_learned_at: number;
    move_learn_method: Ability;
    version_group: Ability;
  }
  interface Gameindex {
    game_index: number;
    version: Ability;
  }
  interface Cries {
    latest: string;
    legacy: string;
  }
  interface AbilityState {
    ability: Ability;
    is_hidden: boolean;
    slot: number;
  }
  interface Ability {
    name: string;
    url: string;
  }